+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Add `gw:characters` command
+- Add Wallet currencies to bank command
 
 ## 1.0.0 - 2021-03-11
 - Refactor to console application